[#50] 파일 정리
